add dataset-loader add --dataset_loader argument
